l-safe/benchmark/requirements.txt
pandas
openai
tqdm
python-dotenv
torch
transformers
accelerate